requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-huggingface
InstructorEmbedding
sentence-transformers
faiss-cpu
